+++ fuzz/fuzz_targets/compare.rs
@@ -4,9 +4,9 @@ use libfuzzer_sys::fuzz_target;
 extern crate zcash_script;
 
 use zcash_script::interpreter::CallbackTransactionSignatureChecker;
-use zcash_script::*;
+use zcash_script::{signature::HashType, *};
 
-fn missing_sighash(_script_code: &[u8], _hash_type: HashType) -> Option<[u8; 32]> {
+fn missing_sighash(_script_code: &[u8], _hash_type: &HashType) -> Option<[u8; 32]> {
     None
 }
 
@@ -34,7 +34,7 @@ fuzz_target!(|tup: (u32, bool, &[u8], &[u8], u32)| {
     );
     assert_eq!(
         ret.0,
-        ret.1.map_err(normalize_error),
+        ret.1.clone().map_err(normalize_error),
         "original Rust result: {:?}",
         ret.1
     );