+++ src/lib.rs
@@ -13,6 +13,7 @@ pub mod interpreter;
 pub mod op;
 mod script;
 pub mod script_error;
+pub mod signature;
 mod zcash_script;
 
 #[cfg(any(test, feature = "test-dependencies"))]
@@ -23,10 +24,10 @@ use std::os::raw::{c_int, c_uint, c_void};
 use tracing::warn;
 
 pub use interpreter::{
-    CallbackTransactionSignatureChecker, DefaultStepEvaluator, HashType, SighashCalculator,
-    SignedOutputs, VerificationFlags,
+    CallbackTransactionSignatureChecker, DefaultStepEvaluator, SighashCalculator, VerificationFlags,
 };
 use script_error::ScriptError;
+use signature::HashType;
 pub use zcash_script::*;
 
 pub struct CxxInterpreter<'a> {
@@ -78,11 +79,17 @@ impl From<cxx::ScriptError> for Error {
                 Error::Ok(ScriptError::UnsatisfiedLockTime)
             }
 
-            cxx::ScriptError_t_SCRIPT_ERR_SIG_HASHTYPE => Error::Ok(ScriptError::SigHashType),
-            cxx::ScriptError_t_SCRIPT_ERR_SIG_DER => Error::Ok(ScriptError::SigDER),
+            cxx::ScriptError_t_SCRIPT_ERR_SIG_HASHTYPE => {
+                Error::Ok(signature::Error::SigHashType(None).into())
+            }
+            cxx::ScriptError_t_SCRIPT_ERR_SIG_DER => {
+                Error::Ok(signature::Error::SigDER(None).into())
+            }
             cxx::ScriptError_t_SCRIPT_ERR_MINIMALDATA => Error::Ok(ScriptError::MinimalData),
             cxx::ScriptError_t_SCRIPT_ERR_SIG_PUSHONLY => Error::Ok(ScriptError::SigPushOnly),
-            cxx::ScriptError_t_SCRIPT_ERR_SIG_HIGH_S => Error::Ok(ScriptError::SigHighS),
+            cxx::ScriptError_t_SCRIPT_ERR_SIG_HIGH_S => {
+                Error::Ok(signature::Error::SigHighS.into())
+            }
             cxx::ScriptError_t_SCRIPT_ERR_SIG_NULLDUMMY => Error::Ok(ScriptError::SigNullDummy),
             cxx::ScriptError_t_SCRIPT_ERR_PUBKEYTYPE => Error::Ok(ScriptError::PubKeyType),
             cxx::ScriptError_t_SCRIPT_ERR_CLEANSTACK => Error::Ok(ScriptError::CleanStack),
@@ -121,7 +128,7 @@ extern "C" fn sighash_callback(
     // And we don’t have access to the flags here to determine if it should be checked.
     if let Some(sighash) = HashType::from_bits(hash_type, false)
         .ok()
-        .and_then(|ht| callback(script_code_vec, ht))
+        .and_then(|ht| callback(script_code_vec, &ht))
     {
         assert_eq!(sighash_out_len, sighash.len().try_into().unwrap());
         // SAFETY: `sighash_out` is a valid buffer created in
@@ -216,9 +223,16 @@ pub fn check_verify_callback<T: ZcashScript, U: ZcashScript>(
 pub fn normalize_error(err: Error) -> Error {
     match err {
         Error::Ok(serr) => Error::Ok(match serr {
+            ScriptError::SignatureEncoding(sig_err) => match sig_err {
+                signature::Error::SigHashType(Some(_)) => {
+                    signature::Error::SigHashType(None).into()
+                }
+                signature::Error::SigDER(Some(_)) => signature::Error::SigDER(None).into(),
+                signature::Error::SigHighS => ScriptError::UnknownError,
+                _ => sig_err.into(),
+            },
             ScriptError::ReadError { .. } => ScriptError::BadOpcode,
             ScriptError::ScriptNumError(_) => ScriptError::UnknownError,
-            ScriptError::SigHighS => ScriptError::UnknownError,
             _ => serr,
         }),
         _ => err,
@@ -280,7 +294,7 @@ impl<T: ZcashScript, U: ZcashScript> ZcashScript for ComparisonInterpreter<T, U>
     ) -> Result<(), Error> {
         let (cxx, rust) =
             check_verify_callback(&self.first, &self.second, script_pub_key, script_sig, flags);
-        if rust.map_err(normalize_error) != cxx {
+        if rust.clone().map_err(normalize_error) != cxx {
             // probably want to distinguish between
             // - one succeeding when the other fails (bad), and
             // - differing error codes (maybe not bad).
@@ -379,17 +393,17 @@ mod tests {
         pub static ref SCRIPT_SIG: Vec<u8> = <Vec<u8>>::from_hex("00483045022100d2ab3e6258fe244fa442cfb38f6cef9ac9a18c54e70b2f508e83fa87e20d040502200eead947521de943831d07a350e45af8e36c2166984a8636f0a8811ff03ed09401473044022013e15d865010c257eef133064ef69a780b4bc7ebe6eda367504e806614f940c3022062fdbc8c2d049f91db2042d6c9771de6f1ef0b3b1fea76c1ab5542e44ed29ed8014c69522103b2cc71d23eb30020a4893982a1e2d352da0d20ee657fa02901c432758909ed8f21029d1e9a9354c0d2aee9ffd0f0cea6c39bbf98c4066cf143115ba2279d0ba7dabe2103e32096b63fd57f3308149d238dcbb24d8d28aad95c0e4e74e3e5e6a11b61bcc453ae").expect("Block bytes are in valid hex representation");
     }
 
-    fn sighash(_script_code: &[u8], _hash_type: HashType) -> Option<[u8; 32]> {
+    fn sighash(_script_code: &[u8], _hash_type: &HashType) -> Option<[u8; 32]> {
         <[u8; 32]>::from_hex("e8c7bdac77f6bb1f3aba2eaa1fada551a9c8b3b5ecd1ef86e6e58a5f1aab952c")
             .ok()
     }
 
-    fn invalid_sighash(_script_code: &[u8], _hash_type: HashType) -> Option<[u8; 32]> {
+    fn invalid_sighash(_script_code: &[u8], _hash_type: &HashType) -> Option<[u8; 32]> {
         <[u8; 32]>::from_hex("08c7bdac77f6bb1f3aba2eaa1fada551a9c8b3b5ecd1ef86e6e58a5f1aab952c")
             .ok()
     }
 
-    fn missing_sighash(_script_code: &[u8], _hash_type: HashType) -> Option<[u8; 32]> {
+    fn missing_sighash(_script_code: &[u8], _hash_type: &HashType) -> Option<[u8; 32]> {
         None
     }
 
@@ -549,7 +563,7 @@ mod tests {
                 &sig[..],
                 flags,
             );
-            prop_assert_eq!(ret.0, ret.1.map_err(normalize_error),
+            prop_assert_eq!(ret.0, ret.1.clone().map_err(normalize_error),
                             "original Rust result: {:?}", ret.1);
         }
 
@@ -584,7 +598,7 @@ mod tests {
                 &sig[..],
                 flags,
             );
-            prop_assert_eq!(ret.0, ret.1.map_err(normalize_error),
+            prop_assert_eq!(ret.0, ret.1.clone().map_err(normalize_error),
                             "original Rust result: {:?}", ret.1);
         }