+++ src/script_error.rs
@@ -1,5 +1,7 @@
 use thiserror::Error;
 
+use crate::signature;
+
 #[derive(Copy, Clone, PartialEq, Eq, Debug, Error)]
 pub enum ScriptNumError {
     #[error("non-minimal encoding of script number")]
@@ -9,11 +11,10 @@ pub enum ScriptNumError {
     Overflow { max_num_size: usize, actual: usize },
 }
 
-#[derive(Copy, Clone, PartialEq, Eq, Debug, Error)]
-#[repr(i32)]
+#[derive(Clone, PartialEq, Eq, Debug, Error)]
 pub enum ScriptError {
     #[error("Ok")]
-    Ok = 0, // Unused (except in converting the C++ error to Rust)
+    Ok, // Unused (except in converting the C++ error to Rust)
 
     #[error("unknown error")]
     UnknownError,
@@ -82,21 +83,14 @@ pub enum ScriptError {
     #[error("unsatisfied locktime condition")]
     UnsatisfiedLockTime,
 
-    // BIP62
-    #[error("signature hash type error")]
-    SigHashType,
-
-    #[error("signature DER encoding error")]
-    SigDER,
+    #[error("signature encoding error: {}", .0)]
+    SignatureEncoding(signature::Error),
 
     #[error("minimal data requirement not met")]
     MinimalData,
 
     #[error("signature push only requirement not met")]
     SigPushOnly,
-
-    #[error("signature s value is too high")]
-    SigHighS,
 
     #[error("signature null dummy error")]
     SigNullDummy,
@@ -111,6 +105,7 @@ pub enum ScriptError {
     #[error("discouraged upgradable NOPs encountered")]
     DiscourageUpgradableNOPs,
 
+    // extensions (these don’t exist in C++, and thus map to `UnknownError`)
     #[error(
         "read error: expected {expected_bytes} bytes, but only {available_bytes} bytes available"
     )]
@@ -127,5 +122,11 @@ pub enum ScriptError {
 impl From<ScriptNumError> for ScriptError {
     fn from(value: ScriptNumError) -> Self {
         ScriptError::ScriptNumError(value)
+    }
+}
+
+impl From<signature::Error> for ScriptError {
+    fn from(value: signature::Error) -> Self {
+        ScriptError::SignatureEncoding(value)
     }
 }