+++ src/test_vectors.rs
@@ -10,6 +10,7 @@ use crate::{
     op::*,
     script::{serialize_num, Opcode},
     script_error::ScriptError,
+    signature,
 };
 use hex::{FromHex, FromHexError};
 
@@ -9434,7 +9435,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(ENDIF),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // 202 CHECKMULTISIGS, fails due to 201 op limit
     TestVector {
@@ -11771,7 +11772,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(NOT),
         ],
         flags: VerificationFlags::StrictEnc,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // 2-of-3 with one valid and one invalid signature due to parse error, nSigs > validSigs
     TestVector {
@@ -11798,7 +11799,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKMULTISIG),
         ],
         flags: DEFAULT_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Increase DERSIG test coverage
 
@@ -11812,7 +11813,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
         ],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Missing S is incorrectly encoded
     TestVector {
@@ -11822,7 +11823,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
         ],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // S with invalid S length is incorrectly encoded
     TestVector {
@@ -11832,7 +11833,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
         ],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Non-integer R is incorrectly encoded
     TestVector {
@@ -11842,7 +11843,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
         ],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Non-integer S is incorrectly encoded
     TestVector {
@@ -11852,21 +11853,21 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
         ],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Zero-length R is incorrectly encoded
     TestVector {
         script_sig: &[H("17"), H("3014020002107777777777777777777777777777777701")],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Zero-length S is incorrectly encoded
     TestVector {
         script_sig: &[H("17"), H("3014021077777777777777777777777777777777020001")],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Negative S is incorrectly encoded
     TestVector {
@@ -11876,7 +11877,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
         ],
         script_pubkey: &[N(0), O(CHECKSIG), O(NOT)],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // Automatically generated test cases
 
@@ -12222,7 +12223,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKSIG),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // P2PK with too much S padding
     TestVector {
@@ -12238,7 +12239,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKSIG),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // P2PK with too little R padding
     TestVector {
@@ -12254,7 +12255,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKSIG),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // P2PK NOT with bad sig with too much R padding
     TestVector {
@@ -12271,7 +12272,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(NOT),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // P2PK NOT with too much R padding
     TestVector {
@@ -12288,7 +12289,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(NOT),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // BIP66 example 1
     TestVector {
@@ -12304,7 +12305,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKSIG),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // BIP66 example 2
     TestVector {
@@ -12321,7 +12322,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(NOT),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // BIP66 example 3
     TestVector {
@@ -12367,7 +12368,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKSIG),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // BIP66 example 6
     TestVector {
@@ -12379,7 +12380,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(NOT),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // BIP66 example 7
     TestVector {
@@ -12404,7 +12405,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKMULTISIG),
         ],
         flags: EMPTY_FLAGS,
-        // FIXME: Should this be returning `SigDER`?
+        // FIXME: Should this be returning `SigDER(None)`?
         result: Err(ScriptError::EvalFalse),
     },
     // BIP66 example 8
@@ -12454,7 +12455,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKMULTISIG),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // BIP66 example 10
     TestVector {
@@ -12477,7 +12478,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(NOT),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // BIP66 example 11
     TestVector {
@@ -12538,7 +12539,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKSIG),
         ],
         flags: EMPTY_FLAGS,
-        result: Err(ScriptError::SigDER),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigDER(None))),
     },
     // P2PK with high S but no LOW_S
     TestVector {
@@ -12793,7 +12794,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(CHECKSIG),
         ],
         flags: VerificationFlags::StrictEnc,
-        result: Err(ScriptError::SigHashType),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigHashType(None))),
     },
     // P2PK NOT with invalid sig and undefined hashtype but no STRICTENC
     TestVector {
@@ -12831,7 +12832,7 @@ pub(crate) const TEST_VECTORS: &[TestVector] = &[
             O(NOT),
         ],
         flags: VerificationFlags::StrictEnc,
-        result: Err(ScriptError::SigHashType),
+        result: Err(ScriptError::SignatureEncoding(signature::Error::SigHashType(None))),
     },
     // 3-of-3 with nonzero dummy but no NULLDUMMY
     TestVector {